+++ container-engine-lib/lib/kurtosis_backend_core/kubernetes_backend_core.go
@@ -3,8 +3,8 @@ package kurtosis_backend_core
 import (
 	"context"
 	"fmt"
-	"github.com/kurtosis-tech/container-engine-lib/lib/k8s_manager"
-	"github.com/kurtosis-tech/container-engine-lib/lib/kurtosis_backend_core/args"
+	"github.com/kurtosis-tech/container-engine-lib/lib/kubernetes_manager"
+	"github.com/kurtosis-tech/kurtosis-engine-server/launcher/args"
 	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
@@ -27,16 +27,17 @@ const (
 type KurtosisKubernetesBackendCore struct {
 	log *logrus.Logger
 
-	k8sManager *k8s_manager.K8sManager
+	kubernetesManager *kubernetes_manager.KubernetesManager
 
 	objAttrsProvider schema.ObjectAttributesProvider
 }
 
-func NewKurtosisKubernetesBackendCore(log *logrus.Logger, k8sManager *k8s_manager.K8sManager, objAttrsProvider schema.ObjectAttributesProvider) *KurtosisKubernetesBackendCore {
+func NewKurtosisKubernetesBackendCore(log *logrus.Logger, k8sManager *kubernetes_manager.KubernetesManager, objAttrsProvider schema.ObjectAttributesProvider) *KurtosisKubernetesBackendCore {
 	return &KurtosisKubernetesBackendCore{
 		log:              log,
-		k8sManager:       k8sManager,
-		objAttrsProvider: objAttrsProvider,
+
+		kubernetesManager: k8sManager,
+		objAttrsProvider:  objAttrsProvider,
 	}
 }
 
@@ -47,6 +48,7 @@ func (kkb KurtosisKubernetesBackendCore) CreateEngine(
 	listenPortNum uint16,
 	engineDataDirpathOnHostMachine string,
 	containerImage string,
+	engineServerArgs *args.EngineServerArgs,
 ) (
 	resultPublicIpAddr net.IP,
 	resultPublicPortNum uint16,
@@ -60,19 +62,13 @@ func (kkb KurtosisKubernetesBackendCore) CreateEngine(
 
 	engineDataDirpathOnHostMachine = defaultHostPathInMinikube
 
-	// getting the args object for the engine server container
-	argsObj, err := args.NewEngineServerArgs(
-		listenPortNum,
-		logLevel.String(),
-		imageVersionTag,
-		engineDataDirpathOnHostMachine,
-	)
+
 	if err != nil {
 		return nil, 0, stacktrace.Propagate(err, "An error occurred creating the engine server args")
 	}
 
 	// getting the envVars for the engine server container
-	envVars, err := args.GetEnvFromArgs(argsObj)
+	envVars, err := args.GetEnvFromArgs(engineServerArgs)
 	if err != nil {
 		return nil, 0, stacktrace.Propagate(err, "An error occurred generating the engine server's environment variables")
 	}
@@ -84,7 +80,7 @@ func (kkb KurtosisKubernetesBackendCore) CreateEngine(
 	)
 
 	// checking if the kurtosis namespace already exists and creating it otherwise
-	kurtosisNamespaceList, err := kkb.k8sManager.GetNamespacesByLabels(ctx,engineLabels)
+	kurtosisNamespaceList, err := kkb.kubernetesManager.GetNamespacesByLabels(ctx,engineLabels)
 	if err != nil {
 		return nil, 0, stacktrace.Propagate(
 			err,
@@ -92,7 +88,7 @@ func (kkb KurtosisKubernetesBackendCore) CreateEngine(
 			engineAttrs.GetLabels())
 	}
 	if len(kurtosisNamespaceList.Items) == 0 {
-		_, err = kkb.k8sManager.CreateNamespace(ctx, kurtosisEngineNamespace, engineLabels)
+		_, err = kkb.kubernetesManager.CreateNamespace(ctx, kurtosisEngineNamespace, engineLabels)
 		if err != nil {
 			return nil, 0, stacktrace.Propagate(
 				err,
@@ -102,7 +98,7 @@ func (kkb KurtosisKubernetesBackendCore) CreateEngine(
 	}
 
 	// creating persistent volume
-	_, err = kkb.k8sManager.CreatePersistentVolume(ctx, engineAttrs.GetName(), engineAttrs.GetLabels(), defaultQuantity, argsObj.EngineDataDirpathOnHostMachine, storageClass)
+	_, err = kkb.kubernetesManager.CreatePersistentVolume(ctx, engineAttrs.GetName(), engineAttrs.GetLabels(), defaultQuantity, engineServerArgs.EngineDataDirpathOnHostMachine, storageClass)
 	if err != nil {
 		return nil, 0, stacktrace.Propagate(
 			err,
@@ -111,7 +107,7 @@ func (kkb KurtosisKubernetesBackendCore) CreateEngine(
 	}
 
 	// creating persistent volume claim
-	_, err = kkb.k8sManager.CreatePersistentVolumeClaim(ctx, kurtosisEngineNamespace, engineAttrs.GetName(), engineAttrs.GetLabels(), defaultQuantity, storageClass)
+	_, err = kkb.kubernetesManager.CreatePersistentVolumeClaim(ctx, kurtosisEngineNamespace, engineAttrs.GetName(), engineAttrs.GetLabels(), defaultQuantity, storageClass)
 	if err != nil {
 		return nil, 0, stacktrace.Propagate(
 			err,
@@ -139,13 +135,13 @@ func (kkb KurtosisKubernetesBackendCore) CreateEngine(
 	}
 
 	// creating deployment
-	_, err = kkb.k8sManager.CreateDeployment(ctx, engineAttrs.GetName(), kurtosisEngineNamespace, engineAttrs.GetLabels(), containerImageAndTag, kurtosisEngineReplicas, volumes, volumeMounts, envVars, engineAttrs.GetName())
+	_, err = kkb.kubernetesManager.CreateDeployment(ctx, engineAttrs.GetName(), kurtosisEngineNamespace, engineAttrs.GetLabels(), containerImageAndTag, kurtosisEngineReplicas, volumes, volumeMounts, envVars, engineAttrs.GetName())
 	if err != nil {
 		return nil, 0, stacktrace.Propagate(err, "An error occurred generating the engine server's environment variables")
 	}
 
 	// creating service
-	service, err := kkb.k8sManager.CreateService(ctx, engineAttrs.GetName(), kurtosisEngineNamespace, engineAttrs.GetLabels(), externalServiceType, int32(listenPortNum), int32(listenPortNum))
+	service, err := kkb.kubernetesManager.CreateService(ctx, engineAttrs.GetName(), kurtosisEngineNamespace, engineAttrs.GetLabels(), externalServiceType, int32(listenPortNum), int32(listenPortNum))
 	if err != nil {
 		return nil, 0, stacktrace.Propagate(err, "An error occurred generating the engine server's environment variables")
 	}
@@ -169,7 +165,7 @@ func (kkb KurtosisKubernetesBackendCore) CreateEngine(
 }
 
 func (kkb KurtosisKubernetesBackendCore) StopEngine(ctx context.Context) error {
-	err := kkb.k8sManager.UpdateDeploymentReplicas(ctx, kurtosisEngineNamespace, engineLabels, int32(zeroReplicas))
+	err := kkb.kubernetesManager.UpdateDeploymentReplicas(ctx, kurtosisEngineNamespace, engineLabels, int32(zeroReplicas))
 	if err != nil {
 		stacktrace.Propagate(err, "an error occurred while trying to stop the engine server with labels %v", engineLabels)
 	}
@@ -178,7 +174,7 @@ func (kkb KurtosisKubernetesBackendCore) StopEngine(ctx context.Context) error {
 }
 
 func (kkb KurtosisKubernetesBackendCore) CleanStoppedEngines(ctx context.Context) ([]string, []error, error) {
-	deploymentsList, err := kkb.k8sManager.GetDeploymentsByLabels(ctx, kurtosisEngineNamespace, engineLabels)
+	deploymentsList, err := kkb.kubernetesManager.GetDeploymentsByLabels(ctx, kurtosisEngineNamespace, engineLabels)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "an error occurred while trying to clean the the stopped engine containers with labels %v", engineLabels)
 	}
@@ -207,22 +203,22 @@ func (kkb KurtosisKubernetesBackendCore) checkIfContainerisRunning(deployment v1
 }
 
 func (kkb KurtosisKubernetesBackendCore) cleanEngineServer(ctx context.Context, name string) error {
-	err := kkb.k8sManager.RemoveDeployment(ctx, kurtosisEngineNamespace, name)
+	err := kkb.kubernetesManager.RemoveDeployment(ctx, kurtosisEngineNamespace, name)
 	if err != nil {
 		return stacktrace.Propagate(err, "an error occurred while trying to delete the deployment from the engine with name %s", name)
 	}
 
-	err = kkb.k8sManager.RemovePersistentVolumeClaim(ctx, kurtosisEngineNamespace, name)
+	err = kkb.kubernetesManager.RemovePersistentVolumeClaim(ctx, kurtosisEngineNamespace, name)
 	if err != nil {
 		return stacktrace.Propagate(err, "an error occurred while trying to delete the persistent volume claim from the engine with name %s", name)
 	}
 
-	err = kkb.k8sManager.RemovePersistentVolume(ctx, name)
+	err = kkb.kubernetesManager.RemovePersistentVolume(ctx, name)
 	if err != nil {
 		return stacktrace.Propagate(err, "an error occurred while trying to delete the persistent volume from the engine with name %s", name)
 	}
 
-	err = kkb.k8sManager.RemoveService(ctx, kurtosisEngineNamespace, name)
+	err = kkb.kubernetesManager.RemoveService(ctx, kurtosisEngineNamespace, name)
 	if err != nil {
 		return stacktrace.Propagate(err, "an error occurred while trying to delete the service from the engine with name %s", name)
 	}
@@ -233,7 +229,7 @@ func (kkb KurtosisKubernetesBackendCore) cleanEngineServer(ctx context.Context, 
 func (kkb KurtosisKubernetesBackendCore) GetEngineStatus(
 	ctx context.Context,
 ) (engineStatus string, ipAddr net.IP, portNum uint16, err error) {
-	deploymentList, err := kkb.k8sManager.GetDeploymentsByLabels(ctx, kurtosisEngineNamespace, engineLabels)
+	deploymentList, err := kkb.kubernetesManager.GetDeploymentsByLabels(ctx, kurtosisEngineNamespace, engineLabels)
 	if err != nil {
 		return "", nil, 0, stacktrace.Propagate(err, "An error occurred getting Kurtosis engine deployments")
 	}
@@ -256,7 +252,7 @@ func (kkb KurtosisKubernetesBackendCore) GetEngineStatus(
 
 	engineDeployment := deployments[0]
 
-	service, err:=kkb.k8sManager.GetServiceByName(ctx, kurtosisEngineNamespace, engineDeployment.Name)
+	service, err:=kkb.kubernetesManager.GetServiceByName(ctx, kurtosisEngineNamespace, engineDeployment.Name)
 
 	publicIpAddr := net.ParseIP(service.Spec.ClusterIP)